WIP Move all business logic into cloud-key-rotator/pkg dir
