+++ pkg/location/locations.go
@@ -1,0 +1,8 @@
+package location
+
+//updatedLocation type
+type UpdatedLocation struct {
+	LocationType string
+	LocationURI  string
+	LocationIDs  []string
+}
