+++ pkg/log/log.go
@@ -1,0 +1,12 @@
+package log
+
+import "go.uber.org/zap"
+
+//stdoutLogger creates a stdout logger
+func StdoutLogger() (logger *zap.Logger) {
+	config := zap.NewProductionConfig()
+	config.OutputPaths = []string{"stdout"}
+	config.ErrorOutputPaths = []string{"stdout"}
+	logger, _ = config.Build()
+	return
+}
